api/requirements.txt
pytest
flask
sqlalchemy
pyyaml
markupsafe

